+++ NEWS.md
@@ -1,5 +1,6 @@
 # glue (development version)
 
+* `glue_sql()` now replaces missing values correctly when collapsing values (#185).
 * `glue_sql()` now always preserves the type of the column even in the presence of missing values (#130)
 
 # glue 1.4.0